+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.dto;
+package nextstep.subway.line.dto;
 
 public class LineRequest {
