+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,9 +1,9 @@
-package nextstep.subway.application;
+package nextstep.subway.station.application;
 
-import nextstep.subway.domain.station.Station;
-import nextstep.subway.domain.station.StationRepository;
-import nextstep.subway.dto.StationRequest;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
+import nextstep.subway.station.dto.StationRequest;
+import nextstep.subway.station.dto.StationResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
