+++ src/main/java/nextstep/subway/station/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain.station;
+package nextstep.subway.station.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
