+++ src/main/java/nextstep/subway/station/dto/StationRequest.java
@@ -1,6 +1,6 @@
-package nextstep.subway.dto;
+package nextstep.subway.station.dto;
 
-import nextstep.subway.domain.station.Station;
+import nextstep.subway.station.domain.Station;
 
 public class StationRequest {
     private String name;
