+++ src/test/java/nextstep/subway/domain/StationNameTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.domain;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
-import nextstep.subway.domain.station.StationName;
+import nextstep.subway.station.domain.StationName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
