+++ src/test/java/nextstep/subway/station/StationAcceptanceTestMethods.java
@@ -9,7 +9,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import java.util.stream.Collectors;
-import nextstep.subway.dto.StationRequest;
+import nextstep.subway.station.dto.StationRequest;
 import org.springframework.http.HttpStatus;
 
 public class StationAcceptanceTestMethods {
